[UPDATE] BuscadorAmbienteCantidad
